requirements.txt
asgiref
crispy-bootstrap5
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
Markdown
MouseInfo
mysql-connector-python
pillow
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
wheel
gunicorn